requirements.txt
beautifulsoup4==4.12.3
cached-property==2.0.1
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.0
cryptography==44.0.0
defusedxml==0.7.1
dnspython==2.7.0
exchangelib==5.5.0
exif==1.6.1
idna==3.10
isodate==0.7.2
loguru==0.7.3
lxml==5.3.0
oauthlib==3.2.2
pillow==11.0.0
plum-py==0.8.7
pycparser==2.22
PyExifTool==0.5.6
Pygments==2.18.0
pyspnego==0.11.2
python-dotenv==1.0.1
requests==2.32.3
requests-oauthlib==2.0.0
requests_ntlm==1.3.0
soupsieve==2.6
tzdata==2024.2
tzlocal==5.2
urllib3==2.2.3
